Stonebreak: Decomposing Game.java
